+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 - added ability to show enemy healthbars only for bosses (#1300)
 - added ability to kill specific enemy types (#1313)
 - added ability to teleport to nearest specific object (#1312)
-- added `/load` command for quick load
+- added `/load` and `/save` commands for even quicker savegame operations
 - added `/give all` variant of the item cheat
 - changed injection files to be placed in its own directory (#1306)
 - changed item cheat sound effects